GRBK-39
Checking in changes to ensure compatibility with GWT 1.6 going forward. Note that the GradebookApplication.launch.tmpl does not 
change. Use the ant build.xml for Hosted mode to get 1.6 to build. You will have to add "rename-to='gradebook'" to the module tag
of GradebookApplication.gxt.xml to get this to work properly.

Note that I'm also uping the GXT version to 1.2.3 here.